Fix syncing extensions from deps.json.

+++ src/engine/src/services/managed-files/deps/update-service.ts
@@ -1,10 +1,12 @@
 import { PrismaClient, SourceNode } from '@prisma/client'
+import { CustomError } from '@/serene-core-server/types/errors'
 import { BuildData } from '@/types/build-types'
 import { ExtensionsData } from '@/types/source-graph-types'
 import { ExtensionMutateService } from '@/services/extensions/extension/mutate-service'
 import { ExtensionQueryService } from '@/services/extensions/extension/query-service'
 import { ProjectSetupService } from '@/services/projects/setup-project'
 import { SourceDepsFileService } from './source-deps-service'
+import { ServerOnlyTypes, VerbosityLevels } from '@/types/server-only-types'
 
 // Services
 const extensionMutateService = new ExtensionMutateService()
@@ -20,12 +22,21 @@ export class DepsUpdateService {
 
   // Code
   checkExtensionInExtensionsData(
-    extensionName: string,
+    extensionId: string,
     extensionsData: ExtensionsData) {
 
+    // Debug
+    const fnName = `${this.clName}.checkExtensionInExtensionsData()`
+
+    // Check every extension in the graph
     for (const extensionNode of extensionsData.extensionNodes) {
 
-      if ((extensionNode.jsonContent as any).id === extensionName) {
+      // Debug
+      // console.log(`${fnName}: checking ` +
+      //   `${(extensionNode.jsonContent as any).id}`)
+
+      // Check
+      if ((extensionNode.jsonContent as any).id === extensionId) {
         return true
       }
     }
@@ -39,26 +50,44 @@ export class DepsUpdateService {
     extensionsData: ExtensionsData,
     depsNodeExtensions: any) {
 
+    // Debug
+    const fnName = `${this.clName}.deleteExtensionsNotInDepsNode()`
+
+    // console.log(`${fnName}: starting with depsNodeExtensions: ` +
+    //   JSON.stringify(depsNodeExtensions))
+
     // Iterate project extensions (from project graph)
     for (const extensionNode of extensionsData.extensionNodes) {
+
+      // Debug
+      console.log(`${fnName}: checking extensionNode..`)
+
+      // Get id
+      const extensionId = (extensionNode.jsonContent as any).id
 
       // Check if extension in depsNode
       var inDepsNode = false
 
-      for (const [name, version] of Object.entries(depsNodeExtensions)) {
+      for (const [id, version] of Object.entries(depsNodeExtensions)) {
 
-        if (this.checkExtensionInExtensionsData(
-          name,
-          extensionsData)) {
-
+        if (id === extensionId) {
           inDepsNode = true
         }
       }
 
       // If extension not in depsNode delete it
-      await extensionMutateService.deleteExtension(
-        prisma,
-        extensionNode.id)
+      if (inDepsNode === false) {
+
+        // Verbose
+        if (ServerOnlyTypes.verbosity >= VerbosityLevels.max) {
+          console.log(`Deleting extension not in deps.json: ${extensionId}`)
+        }
+
+        // Delete
+        await extensionMutateService.deleteExtension(
+          prisma,
+          extensionNode.id)
+      }
     }
   }
 
@@ -68,6 +97,11 @@ export class DepsUpdateService {
     extensionsData: ExtensionsData,
     depsNodeExtensions: any) {
 
+    // Debug
+    const fnName = `${this.clName}.loadExtensionsFromDepsNode()`
+
+    // console.log(`${fnName}: starting..`)
+
     // Validate
     if (Array.isArray(depsNodeExtensions)) {
 
@@ -76,17 +110,24 @@ export class DepsUpdateService {
     }
 
     // Iterate depsNode extensions
-    for (const [name, version] of Object.entries(depsNodeExtensions)) {
+    for (const [id, version] of Object.entries(depsNodeExtensions)) {
 
+      // Debug
+      // console.log(`${fnName}: checking if ${id} is loaded..`)
+
+      // Check if the extension exists in the graph
       if (!this.checkExtensionInExtensionsData(
-        name,
+        id,
         extensionsData)) {
+
+        // Debug
+        // console.log(`${fnName}: loading extensions ${id}..`)
 
         // Load extension into project
         await extensionMutateService.loadExtensionsInSystemToUserProject(
           prisma,
           projectNode.instanceId,
-          [name])
+          [id])
       }
     }
   }
@@ -120,6 +161,9 @@ export class DepsUpdateService {
     projectNode: SourceNode,
     depsNode: SourceNode) {
 
+    // Debug
+    const fnName = `${this.clName}.updateExtensions()`
+
     // Get extensions in the project
     const projectExtensionsData = await
       extensionQueryService.loadExtensions(
@@ -135,14 +179,24 @@ export class DepsUpdateService {
       process.exit(1)
     }
 
+    // Debug
+    if (ServerOnlyTypes.verbosity >= VerbosityLevels.max) {
+
+      console.log(`${fnName}: projectExtensionsData.extensionNodes: ` +
+        `${projectExtensionsData.extensionNodes.length}`)
+    }
+
+    // Get depsNode extensions
+    const depsNodeExtensions = (depsNode?.jsonContent as any)?.extensions
+
     // Try to load any extensions not in the project
-    if ((depsNode?.jsonContent as any)?.extensions != null) {
+    if (depsNodeExtensions != null) {
 
       await this.loadExtensionsFromDepsNode(
         prisma,
         projectNode,
         projectExtensionsData,
-        (depsNode?.jsonContent as any)?.extensions)
+        depsNodeExtensions)
     }
 
     // Try to delete any extensions not in the new depsNode
@@ -150,6 +204,6 @@ export class DepsUpdateService {
       prisma,
       projectNode,
       projectExtensionsData,
-      depsNode)
+      depsNodeExtensions)
   }
 }
